requirements.txt
# Deep Learning
git+https://github.com/meteofrance/mfai@v5.1.1
lightning==2.5.1
tensorboard==2.16.2
networkx==3.3
tensorflow_hub==0.16.1
torchmetrics==1.8.2
torchvision==0.21.0+cu118
transformers==4.57.1
langchain==0.2.6
sentence-transformers==2.7.0
chromadb==0.4.24
jsonargparse[signatures]==4.32.1
pytorch-msssim==1.0.0
flash-attn==2.8.3
trl==0.22.0
peft==0.17.1
onnxruntime==1.23.1

# Plots
cartopy==0.24.1
matplotlib==3.8.4
shapely==2.0.3 --no-binary=shapely
gif==23.3.0
opencv-python-headless==4.9.0.80
seaborn==0.13.2

# Data Science
scikit-learn==1.4.2
scikit-image==0.23.*
skl2onnx==1.17.0
skops==0.9.0
geopandas==0.14.3
xarray==2025.3.1
einops==0.8.1

# Meteorology
ecmwf-api-client==1.6.3
cfgrib==0.9.13.0
netCDF4==1.7.2
dask==2025.1.0
eccodes==2.41.0

# Python
numpy==1.26.4
notebook==7.1.3
py-spy==0.3.14
filprofiler==2023.3.1
icecream==2.1.3
joblib==1.4.0
argparse-dataclass==2.0.0
tqdm==4.66.4
pyment==0.3.3
protobuf==3.20.*
Cython==3.0.10
PyYAML==6.0.1
psutil==6.0.0
pypdf==4.2.0
lxml==5.2.1
pytest==8.2.2
pyright[nodejs]==1.1.407

# Others
flask==3.0.3
opencv-python==4.11.0.86
pillow==11.0.0

# Metaibricks
pydantic==2.10.3
pydantic-settings==2.6.1
paramiko==3.5.0